+++ TP3/src/Solution.h
@@ -18,26 +18,26 @@ public:
   ~Solution();
   void setNbPoints(int nbPoints);
   void setNbIncidents(std::vector<int>& nbIncidents);
-  void setSentier(int depart, int arrivee, float cout);
-  void setSentiers(std::vector<std::vector<float> >& sentiers);
+  void setSentier(int depart, int arrivee, double cout);
+  void setSentiers(std::vector<std::vector<double> >& sentiers);
   void deleteSentier(int depart, int arrivee);
-  void addTabou(float coutTotal);
+  void addTabou(double coutTotal);
   int getNbPoints();
   std::vector<int> getNbIncidents();
-  std::vector<std::vector<float> > getSentiers();
-  float getCoutTotal();
-  std::deque<float> getTabou();
+  std::vector<std::vector<double> > getSentiers();
+  double getCoutTotal();
+  std::deque<double> getTabou();
   std::pair<Erreur, int> verifier(Exemplaire& e);
   void afficher();
-  float calculer();
-  bool estTabou(float coutTotal);
+  double calculer();
+  bool estTabou(double coutTotal);
 
 private:
   int nbPoints;
   std::vector<int> nbIncidents;
-  std::vector<std::vector<float> > sentiers;
-  float coutTotal;
-  std::deque<float> tabou;
+  std::vector<std::vector<double> > sentiers;
+  double coutTotal;
+  std::deque<double> tabou;
 };
 
 #endif
